+++ packages/fastboard-core/src/behaviors/index.ts
@@ -50,3 +50,13 @@ for (const kind in DefaultApps) {
 }
 
 export const register = WindowManager.register.bind(WindowManager);
+
+declare let __NAME__: string, __VERSION__: string;
+
+export const version = __VERSION__;
+
+if (typeof window !== "undefined") {
+  let str = (window as { __netlessUA?: string }).__netlessUA || "";
+  str += ` ${__NAME__}@${version} `;
+  (window as { __netlessUA?: string }).__netlessUA = str;
+}