+++ packages/fastboard-react/vite.config.ts
@@ -5,4 +5,5 @@ import react from "@vitejs/plugin-react";
 export default defineConfig({
   plugins: [react(), svelte()],
   envDir: "../..",
+  define: { __NAME__: '"@netless/fastboard"', __VERSION__: '"develop"' },
 });