+++ packages/fastboard-ui/vite.config.ts
@@ -3,4 +3,6 @@ import { svelte } from "@sveltejs/vite-plugin-svelte";
 
 export default defineConfig({
   plugins: [svelte({ experimental: { useVitePreprocess: true } })],
+  envDir: "../..",
+  define: { __NAME__: '"@netless/fastboard"', __VERSION__: '"develop"' },
 });